added email param check

+++ CLOCKEDIN_Backend/views/admin/views.py
@@ -7,8 +7,10 @@ from rest_framework.views import APIView
 from CLOCKEDIN_Backend.models import Company, Invitation
 from CLOCKEDIN_Backend.permissions import IsAdmin, IsManager
 from CLOCKEDIN_Backend.serializers import InviteSerializer
-from CLOCKEDIN_Backend.utils.mailing.canceled_invitation_mail import send_cancelled_invitation_email
-from CLOCKEDIN_Backend.utils.mailing.welcome_mail_sender import send_welcome_email
+from CLOCKEDIN_Backend.utils.mailing.canceled_invitation_mail import \
+    send_cancelled_invitation_email
+from CLOCKEDIN_Backend.utils.mailing.welcome_mail_sender import \
+    send_welcome_email
 
 
 class InviteView(APIView):
@@ -46,18 +48,21 @@ class InviteView(APIView):
             {"message": f"Invitation sent and user created! {request.user.is_authenticated}, {request.user}"},
             status=status.HTTP_200_OK,
         )
-
+    
     def delete(self, request):
         # Validate company ID
         company_id = request.user.company_id
         if not company_id:
             return Response({"error": "Company ID not in inviter data"}, status=status.HTTP_400_BAD_REQUEST)
-
+        
         # Get email value
         params = request.query_params
-        inv_email = params.get("email")  # Dont wheter to check if the email exists
+        inv_email = params.get("email") #Dont wheter to check if the email exists
 
         # Get invitation related to that email and company
+        if not inv_email:
+            return Response({"error": "Email not provided in the parameters"}, status=status.HTTP_400_BAD_REQUEST)
+        
         invitation = get_object_or_404(Invitation, email=inv_email, company=company_id, status="pending")
         invitation.delete()
 
@@ -67,3 +72,9 @@ class InviteView(APIView):
             {"message": f"Invitation has been cancelled! {request.user.is_authenticated}, {request.user}"},
             status=status.HTTP_200_OK,
         )
+
+
+
+        
+
+
